Update existing test coverage
